refac: Split pytest default fixtures from helpers
